added back exoplanet modal for less data fetching
